Cleans master branch up of no longer needed files
