--- config.js
@@ -1,5 +1,0 @@
-module.exports = {
-    consumer_key:         'AD5TEbpI9AsOZfCWeEh7VPfFv',
-    consumer_secret:      'DwzVgTRnzNFwayHj5b9Hd9DBOsDioPdqnQpmJDDvaCIlIvXwpZ',
-    app_only_auth:        true
-}
