--- package.json
@@ -1,27 +1,0 @@
-{
-  "name": "Project-1",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "lite-server"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/asanchez3315/Project-1.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/asanchez3315/Project-1/issues"
-  },
-  "homepage": "https://github.com/asanchez3315/Project-1#readme",
-  "devDependencies": {
-    "lite-server": "^2.4.0"
-  },
-  "dependencies": {
-    "moment": "^2.22.2",
-    "twit": "^2.2.11"
-  }
-}
