--- query-test.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <title>Single AJAX</title>
-</head>
-
-<body>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script type="text/javascript">
-
-    $.ajax({
-    type: "POST",
-    url: 'https://cors-anywhere.herokuapp.com/http://text-processing.com/api/sentiment',
-    text: 'This tweet is sad I hope i feel better soon'
-}).then(function(response) {
-      console.log(response);
-    });
-// ---------------------------------------------------------
-  </script>
-
-</body>
-
-</html>
-
-
-)
